requirements-test.txt
# test runs requirements (versions we'll be testing against) - automatically updated by requires.io
coverage==4.2 # pytest-cov
pylama==7.2.3
pyflakes==1.3.0
pydocstyle==1.1.1
pycodestyle==2.2.0
mccabe==0.5.2
pyroma==2.2
pytest==3.0.4
rabbitpy==1.0.0
path.py==8.2.1
port-for==0.3.1
mirakuru==0.8.2
